Make the option validator and the numeric readers agree
Review findings on the numeric-coercion change. Three held up under
measurement; two did not, and are recorded here rather than acted on.

A go feature option of a DEFINED numeric type (`type Retries int`) was
accepted by the option validator, which classifies by reflect.Kind, and
then dropped by the reader, whose type switch matches exact dynamic
types. Measured: validation passes, foptInt returns the default. The
readers now fall back to reflect for that case, so both halves agree on
what counts as a number.

`json.Number` was the mirror image: the readers honour it, the validator
calls it a string because that is its reflect.Kind, and makeOptions
DISCARDS the validation error, so the value survived and the coercion
worked. Latent rather than broken, but an option judged by a rule
nothing else applies is the same defect one layer up. makeOptions now
canonicalises json.Number ahead of validation.

csharp `decimal -> int` is the one conversion C# range-checks even
unchecked, so an oversized decimal option threw where every other target
coerces. Routed via double, which is what the neighbouring arm does.

The mint-form guard covered 15 of the 20 SDK targets and SKIPPED a
target whose file it could not find. It now carries csharp, elixir,
ocaml, rust and zig — each expressing the same id form in its own
syntax — fails on a missing file, and compares its own coverage against
the shipped target list so a new target cannot slip past it.

Two findings did not survive checking, and nothing changed for them:
csharp validation does not reject the wider numeric family, because an
unrecognised type maps to T.Any, an all-bits mask that satisfies every
type test; and the go json.Number branch IS reached in the production
path, per the measurement above.

Each fix was reverted and seen to fail its own assertion: the defined
type drops to the default budget, the uncanonicalised option arrives as
json.Number, a drifted rust mint form is named, a moved csharp file is
named. The go cases run inside a generated SDK, through the real
MakeOptions, closing the gap left by a harness that called Init directly.

Full suite: 1385 pass, 0 fail, 13 skipped. Golden moved for the three
templates touched. csharp remains unbuilt: no dotnet on this machine.

+++ ts/project/.sdk/tm/csharp/feature/FeatureOptions.cs
@@ -34,7 +34,9 @@ internal static class FeatureOptions
             byte n => n,
             double n => (int)n,
             float n => (int)n,
-            decimal n => (int)n,
+            // Via double: decimal->int is the one conversion C# range-checks
+            // even unchecked, and no other target throws on an oversized option.
+            decimal n => (int)(double)n,
             _ => def,
         };
     }
